+++ src/backends/db2/standard-use-type.cpp
@@ -37,12 +37,12 @@ void *db2_standard_use_type_backend::prepare_for_bind(
     case x_long_long:
         sqlType = SQL_BIGINT;
         cType = SQL_C_SBIGINT;
-        size = sizeof(long long);
+        size = sizeof(int64_t);
         break;
     case x_unsigned_long_long:
         sqlType = SQL_BIGINT;
         cType = SQL_C_UBIGINT;
-        size = sizeof(unsigned long long);
+        size = sizeof(uint64_t);
         break;
     case x_double:
         sqlType = SQL_DOUBLE;
