+++ src/backends/sqlite3/vector-into-type.cpp
@@ -288,10 +288,10 @@ void sqlite3_vector_into_type_backend::resize(std::size_t sz)
         resize_vector<int>(data_, sz);
         break;
     case x_long_long:
-        resize_vector<long long>(data_, sz);
+        resize_vector<int64_t>(data_, sz);
         break;
     case x_unsigned_long_long:
-        resize_vector<unsigned long long>(data_, sz);
+        resize_vector<uint64_t>(data_, sz);
         break;
     case x_double:
         resize_vector<double>(data_, sz);
@@ -326,10 +326,10 @@ std::size_t sqlite3_vector_into_type_backend::size()
         sz = get_vector_size<int>(data_);
         break;
     case x_long_long:
-        sz = get_vector_size<long long>(data_);
+        sz = get_vector_size<int64_t>(data_);
         break;
     case x_unsigned_long_long:
-        sz = get_vector_size<unsigned long long>(data_);
+        sz = get_vector_size<uint64_t>(data_);
         break;
     case x_double:
         sz = get_vector_size<double>(data_);
