+++ tests/empty/test-empty.cpp
@@ -78,12 +78,12 @@ TEST_CASE("Dummy test", "[empty]")
     sql << query, into(i);
 
 #if defined (__LP64__) || ( __WORDSIZE == 64 )
-    long int li = 9;
+    int64_t li = 9;
     sql << "insert", use(li);
     sql << "select", into(li);
 #endif
 /*
-    long long ll = 11;
+    int64_t ll = 11;
     sql << "insert", use(ll);
     sql << "select", into(ll);
 */
